pyproject.toml
[project]
name = "sensor-dashboard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arel>=0.3.0",
    "fastapi>=0.116.1",
    "jinja2>=3.1.6",
    "matplotlib>=3.10.8",
    "pyserial>=3.5",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "ruff>=0.12.7",
    "sse-starlette>=3.0.2",
    "ty>=0.0.1a16",
    "uvicorn>=0.35.0",
]
